user_app/requirements.txt
flask
requests
pybluez2
paho-mqtt
Flask-MQTT
APScheduler
Flask-APScheduler
